bot/requirements.txt
#taken from pip freeze
discord.py==2.3.1
python-dotenv==1.0.0
requests==2.31.0
uuid==1.30